Ajuste DTO e service

+++ dev/back/pagamentos/src/main/java/br/com/mobiauto/pagamentos/service/PagamentoService.java
@@ -56,6 +56,9 @@ public class PagamentoService {
 
 	@Transactional
     public void excluirPagamento(Long id) {
+		if (!pagamentoRepository.existsById(id)) {
+            throw new EntityNotFoundException("Pagamento não encontrado");
+        }
 		pagamentoRepository.deleteById(id);
     }
